Stop doing unit conversion for temperature readings
